finish booking screen

+++ lib/screens/booking_screen.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:get/get.dart';
+import 'package:intl/intl.dart' as intl;
 import 'package:restaurant_system/models/booking_model.dart';
 import 'package:restaurant_system/networks/rest_api.dart';
 import 'package:restaurant_system/screens/add_booking_screen.dart';
+import 'package:restaurant_system/screens/order_screen.dart';
+import 'package:restaurant_system/utils/enums/enum_order_type.dart';
+import 'package:restaurant_system/utils/utils.dart';
 
 class BookingScreen extends StatefulWidget {
   const BookingScreen({Key? key}) : super(key: key);
@@ -14,6 +18,7 @@ class BookingScreen extends StatefulWidget {
 
 class _BookingScreenState extends State<BookingScreen> {
   List<BookingModel> booking = [];
+  final String dateFormat = 'yyyy-MM-dd HH:mm';
 
   @override
   void initState() {
@@ -53,21 +58,45 @@ class _BookingScreenState extends State<BookingScreen> {
           subtitle: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              Text('${'Date'.tr} : ${booking[index].bookingDate}'),
+              Text('${'Date'.tr} : ${booking[index].bookingDate.isEmpty ? "" : intl.DateFormat(dateFormat).format(DateTime.parse(booking[index].bookingDate))}'),
               Text('${'Hours'.tr} : ${booking[index].noOfHours}'),
               Text('${'Persons'.tr} : ${booking[index].noOfPersons}'),
             ],
           ),
-          trailing: IconButton(
-            onPressed: () async {
-              var result = await RestApi.deleteBooking(
-                id: booking[index].id,
-              );
-              if (result) {
-                getBooking();
-              }
-            },
-            icon: const Icon(Icons.delete),
+          trailing: Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              IconButton(
+                onPressed: () async {
+                  var resultAreYouSure = await Utils.showAreYouSureDialog(title: 'Finish');
+                  if (resultAreYouSure) {
+                    var result = await RestApi.deleteBooking(
+                      id: booking[index].id,
+                    );
+                    if (result) {
+                      Get.back();
+                      Get.to(() => const OrderScreen(type: OrderType.takeAway));
+                    }
+                  }
+                },
+                icon: const Icon(Icons.check),
+              ),
+              const SizedBox(width: 10),
+              IconButton(
+                onPressed: () async {
+                  var resultAreYouSure = await Utils.showAreYouSureDialog(title: 'Delete');
+                  if (resultAreYouSure) {
+                    var result = await RestApi.deleteBooking(
+                      id: booking[index].id,
+                    );
+                    if (result) {
+                      getBooking();
+                    }
+                  }
+                },
+                icon: const Icon(Icons.delete),
+              )
+            ],
           ),
         ),
       ),
